requirements.txt
python-dotenv==1.0.1
pydantic>=2.9.2
pytest==8.0.2
google-generativeai==0.8.4
